Removed warnings from visual c net
git-svn-id: file:///fltk/svn/trunk@5812 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fltk/AnsiWidget.h
@@ -59,7 +59,7 @@ class AnsiWidget : public Widget {
     int textHeight(void);
     int getWidth()  {return w();}
     int getHeight() {return h();}
-    void setFontSize(int i) {labelsize(i);}
+    void setFontSize(float i) {labelsize(i);}
     int getFontSize() {return (int)labelsize();}
     void beep() const;
     static Color ansiToFltk(long color);
